+++ 01-WEATHERAPP/src/App.css
@@ -4,7 +4,7 @@
   text-align: center;
 }
 
-.body{
+body{
   width: 100%;
   height: 100%;
   background-image: url('assets/img/nubes.jpg');
